# models/devlogCategory.js
const { DataTypes } = require('sequelize');
const sequelize = require('../config/database');

const DevlogCategory = sequelize.define('DevlogCategory', {
  id: {
    type: DataTypes.INTEGER,
    autoIncrement: true,
    primaryKey: true
  },
  name: {
    type: DataTypes.STRING,
    allowNull: false,
    validate: {
      notEmpty: { msg: 'Name is required' }
    }
  },
  slug: {
    type: DataTypes.STRING,
    allowNull: false,
    unique: true,
    validate: {
      notEmpty: { msg: 'Slug is required' }
    }
  },
  description: {
    type: DataTypes.TEXT,
    allowNull: true
  }
}, {
  tableName: 'devlog_categories',
  timestamps: true
});

module.exports = DevlogCategory;