# openapi.yaml
openapi: 3.0.3
info:
  title: Hi Planet API
  description: API documentation for the Hi Planet Backend, including User Authentication and Tester Applications.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development server
paths:
  /auth/register:
    post:
      summary: Register a new user
      description: Create a new administrator/team member account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                  example: adminuser
                email:
                  type: string
                  format: email
                  example: admin@hiplanet.com
                password:
                  type: string
                  format: password
                  example: password123
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: User registered successfully
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      token:
                        type: string
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          description: Validation error or duplicate username/email
  /auth/login:
    post:
      summary: Log in a user
      description: Authenticate credentials (email or username) and retrieve a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - identifier
                - password
              properties:
                identifier:
                  type: string
                  description: Email or Username
                  example: admin@hiplanet.com
                password:
                  type: string
                  format: password
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Login successful
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      token:
                        type: string
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          description: Invalid credentials
  /users:
    get:
      summary: Get all users
      description: Retrieve a list of registered team members. Protected by JWT.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User list fetched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
    post:
      summary: Create a user
      description: Create a user from admin panel. Protected by JWT.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                  example: testuser
                email:
                  type: string
                  format: email
                  example: test@hiplanet.com
                password:
                  type: string
                  format: password
                  example: testpassword123
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: User created successfully
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
  /tester-applications:
    post:
      summary: Submit a tester application
      description: Public endpoint for potential testers to submit applications.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - name
                - platform
                - why
                - nda
              properties:
                email:
                  type: string
                  format: email
                  example: tester@example.com
                name:
                  type: string
                  example: SuperPlayer
                platform:
                  type: string
                  enum: [Windows, macOS, Both]
                  example: Windows
                why:
                  type: string
                  example: I have played 500 hours of horde-roguelite games.
                country:
                  type: string
                  example: South Korea
                nda:
                  type: boolean
                  example: true
      responses:
        '201':
          description: Application submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Application submitted successfully
                  data:
                    $ref: '#/components/schemas/TesterApplication'
        '400':
          description: Validation error
    get:
      summary: List all tester applications
      description: Retrieve all beta test applications. Protected by JWT. Supports pagination.
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            default: 1
            example: 1
        - name: limit
          in: query
          description: Maximum number of applications to return
          required: false
          schema:
            type: integer
            default: 10
            example: 10
      responses:
        '200':
          description: List of applications retrieved successfully with pagination details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TesterApplication'
                  pagination:
                    type: object
                    properties:
                      totalItems:
                        type: integer
                        example: 25
                      totalPages:
                        type: integer
                        example: 3
                      currentPage:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 10
        '401':
          description: Unauthorized
  /tester-applications/{id}:
    put:
      summary: Update a tester application
      description: Update tester details or application status (pending, approved, rejected). Protected by JWT.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                name:
                  type: string
                platform:
                  type: string
                  enum: [Windows, macOS, Both]
                why:
                  type: string
                country:
                  type: string
                nda:
                  type: boolean
                status:
                  type: string
                  enum: [pending, approved, rejected]
                  example: approved
      responses:
        '200':
          description: Application updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Application updated successfully
                  data:
                    $ref: '#/components/schemas/TesterApplication'
        '401':
          description: Unauthorized
        '404':
          description: Application not found
    delete:
      summary: Delete a tester application
      description: Remove an application from the database. Protected by JWT.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Application deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Application deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Application not found
  /econews-blogs:
    get:
      summary: Get all blog posts
      description: Retrieve a list of blog posts. Public access. Supports searching, filtering, and pagination.
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            default: 1
            example: 1
        - name: limit
          in: query
          description: Maximum number of blog posts to return
          required: false
          schema:
            type: integer
            default: 10
            example: 10
        - name: category
          in: query
          description: Filter posts by category
          required: false
          schema:
            type: string
            example: Nature
        - name: search
          in: query
          description: Search term matching title, description, or author
          required: false
          schema:
            type: string
            example: Forest
        - name: sortBy
          in: query
          description: Database column to sort by
          required: false
          schema:
            type: string
            enum: [title, author, category, date_created, createdAt]
            default: date_created
        - name: sortOrder
          in: query
          description: Sorting direction
          required: false
          schema:
            type: string
            enum: [asc, desc, ASC, DESC]
            default: desc
      responses:
        '200':
          description: List of blog posts retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EcoNewsBlog'
                  pagination:
                    type: object
                    properties:
                      totalItems:
                        type: integer
                        example: 5
                      totalPages:
                        type: integer
                        example: 1
                      currentPage:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 10
    post:
      summary: Create a new blog post
      description: Add a new blog post with an optional image upload. Protected by JWT.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - title
                - author
                - description
                - category
                - slug
              properties:
                title:
                  type: string
                  example: Planting Trees for the Future
                author:
                  type: string
                  example: Sean Beck
                description:
                  type: string
                  description: Blog body, supports rich-text HTML markup.
                  example: <h1>Planting Trees</h1><p>Let's make our planet greener!</p>
                category:
                  type: string
                  example: Nature
                slug:
                  type: string
                  example: planting-trees-for-the-future
                image:
                  type: string
                  format: binary
                  description: Image file to upload (JPEG, PNG, GIF, WEBP)
                date_created:
                  type: string
                  format: date-time
                  description: Custom creation date. Defaults to current date if omitted.
                  example: '2026-06-04T12:00:00.000Z'
      responses:
        '201':
          description: Blog post created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Blog post created successfully
                  data:
                    $ref: '#/components/schemas/EcoNewsBlog'
        '400':
          description: Validation or file size error
        '401':
          description: Unauthorized
  /econews-blogs/{id}:
    get:
      summary: Get a blog post by ID
      description: Retrieve a single blog post. Public access.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Blog post details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/EcoNewsBlog'
        '404':
          description: Blog post not found
    put:
      summary: Update a blog post
      description: Update blog details or replace the image. Protected by JWT.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Updated Planting Trees title
                author:
                  type: string
                  example: Sean Beck
                description:
                  type: string
                  example: Updated content
                category:
                  type: string
                  example: Action
                slug:
                  type: string
                  example: updated-planting-trees
                image:
                  type: string
                  format: binary
                date_created:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Blog post updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Blog post updated successfully
                  data:
                    $ref: '#/components/schemas/EcoNewsBlog'
        '400':
          description: Validation or upload error
        '401':
          description: Unauthorized
        '404':
          description: Blog post not found
    delete:
      summary: Delete a blog post
      description: Remove a blog post and its associated image from the server. Protected by JWT.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Blog post deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Blog post deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Blog post not found
  /devlog-categories:
    get:
      summary: Get all devlog categories
      description: Retrieve a list of devlog categories. Public access.
      responses:
        '200':
          description: List of categories retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DevlogCategory'
    post:
      summary: Create a new devlog category
      description: Add a new devlog category. Protected by JWT.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - slug
              properties:
                name:
                  type: string
                  example: General
                slug:
                  type: string
                  example: general
                description:
                  type: string
                  example: General devlog updates
      responses:
        '201':
          description: Devlog category created successfully
        '400':
          description: Validation error
        '401':
          description: Unauthorized
  /devlog-categories/{id}:
    get:
      summary: Get a devlog category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Devlog category retrieved
        '404':
          description: Not found
    put:
      summary: Update a devlog category
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                slug:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Updated successfully
    delete:
      summary: Delete a devlog category
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted successfully
  /devlog-blogs:
    get:
      summary: Get all devlog posts
      description: Retrieve a list of devlog posts. Public access. Supports searching, filtering by category_id, and pagination.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: category_id
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of devlog posts retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DevlogBlog'
    post:
      summary: Create a new devlog post
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - title
                - author
                - description
                - category_id
                - slug
              properties:
                title:
                  type: string
                author:
                  type: string
                description:
                  type: string
                category_id:
                  type: integer
                slug:
                  type: string
                image:
                  type: string
                  format: binary
                date_created:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Created successfully
  /devlog-blogs/{id}:
    get:
      summary: Get a devlog post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Devlog post details
    put:
      summary: Update a devlog post
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                author:
                  type: string
                description:
                  type: string
                category_id:
                  type: integer
                slug:
                  type: string
                image:
                  type: string
                  format: binary
                date_created:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Updated successfully
    delete:
      summary: Delete a devlog post
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted successfully
  /devlog-blogs/slug/{slug}:
    get:
      summary: Get a devlog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Devlog post details
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: adminuser
        email:
          type: string
          format: email
          example: admin@hiplanet.com
        createdAt:
          type: string
          format: date-time
          example: '2026-05-29T10:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2026-05-29T10:00:00.000Z'
    TesterApplication:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: tester@example.com
        name:
          type: string
          example: SuperPlayer
        platform:
          type: string
          example: Windows
        why:
          type: string
          example: I love roguelites!
        country:
          type: string
          example: South Korea
        nda:
          type: boolean
          example: true
        status:
          type: string
          example: pending
        createdAt:
          type: string
          format: date-time
          example: '2026-05-29T10:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2026-05-29T10:00:00.000Z'
    EcoNewsBlog:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Planting Trees for the Future
        author:
          type: string
          example: Sean Beck
        description:
          type: string
          example: <h1>Planting Trees</h1><p>Let's make our planet greener!</p>
        category:
          type: string
          example: Nature
        slug:
          type: string
          example: planting-trees-for-the-future
        image:
          type: string
          nullable: true
          example: /uploads/forest-123456789.png
        date_created:
          type: string
          format: date-time
          example: '2026-06-04T12:00:00.000Z'
        createdAt:
          type: string
          format: date-time
          example: '2026-06-04T12:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2026-06-04T12:00:00.000Z'
    DevlogCategory:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: General
        slug:
          type: string
          example: general
        description:
          type: string
          example: General devlog updates
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DevlogBlog:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Devlog 1
        author:
          type: string
          example: Admin
        description:
          type: string
          example: <p>First update</p>
        category_id:
          type: integer
          example: 1
        slug:
          type: string
          example: devlog-1
        image:
          type: string
          nullable: true
          example: /uploads/image.png
        date_created:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time